refactor: restaurantList stores 폴더 하위로 이동
